% Auto-generated: do not edit by hand
\name{ldLdInput}

\alias{ldLdInput}

\title{LdInput component}

\description{

}

\usage{
ldLdInput(label=NULL, updateValue=NULL, value=NULL)
}

\arguments{
\item{label}{Character. }

\item{updateValue}{}

\item{value}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

