% Auto-generated: do not edit by hand
\name{ldLiquidPlotly}

\alias{ldLiquidPlotly}

\title{LiquidPlotly component}

\description{
Component description
}

\usage{
ldLiquidPlotly(id=NULL, label=NULL, value=NULL)
}

\arguments{
\item{id}{Character. Unique ID to identify this component in Dash callbacks.}

\item{label}{Character. }

\item{value}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

